Potential fix for crashing gd_oo_tests (SF.net #3518190).
The IMG_PTR needed to be a longlong, but was a long.  At that made all the difference.
I do not have BSD AMD64, but was getting a segfault after "ok 19" when the DESTROY
was called.  I have not removed the TODO from the gd_oo_tests.t, we should do that
after it is certain that this does not break anything for anybody else.